(Chore:) Refactoring
